requirements.txt
ollama>=0.1.0
requests>=2.28.0
pyinstaller>=5.0.0
datetime>=4.0.0
openai>=1.0.0
anthropic>=0.5.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
